pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homectl"
version = "0.1.0"
description = "CLI for managing a home-server hosting platform backed by Cloudflare Tunnel, Traefik, and Docker Compose."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Jinja2>=3.1.0",
  "PyYAML>=6.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
homectl = "homectl.main:run"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["homectl*"]

[tool.setuptools.package-data]
homectl = [
  "py.typed",
  "templates/**/*.j2",
  "templates/**/*.md.j2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
